Remove itemRenderer from styles.
