--- assets/src/edit-story/components/panels/design/preset/presets.js
@@ -1,59 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-
-/**
- * Internal dependencies
- */
-import PresetGroup from './presetGroup';
-
-function Presets({
-  presets,
-  handleOnClick,
-  itemRenderer,
-  isEditMode,
-  type,
-  handleAddPreset,
-  isLocal = false,
-}) {
-  return (
-    <PresetGroup
-      itemRenderer={itemRenderer}
-      presets={presets}
-      type={type}
-      isEditMode={isEditMode}
-      handleClick={handleOnClick}
-      handleAddPreset={handleAddPreset}
-      isLocal={isLocal}
-    />
-  );
-}
-
-Presets.propTypes = {
-  presets: PropTypes.array.isRequired,
-  itemRenderer: PropTypes.func.isRequired,
-  type: PropTypes.string.isRequired,
-  handleOnClick: PropTypes.func.isRequired,
-  isEditMode: PropTypes.bool.isRequired,
-  handleAddPreset: PropTypes.func.isRequired,
-  isLocal: PropTypes.bool,
-};
-
-export default Presets;
