+++ assets/src/edit-story/components/panels/design/preset/test/presetGroup.js
@@ -57,7 +57,7 @@ const getThirdArguments = (calls) => {
   return [...calls.map((args) => args[2])];
 };
 
-describe('globalStoryStyles/PresetGroup', () => {
+describe('stylePresets/PresetGroup', () => {
   it('should modify active index correctly with arrow keys', () => {
     const { itemRenderer, getAllByRole } = setupPresetGroup();
     expect(getThirdArguments(itemRenderer.mock.calls)).toStrictEqual([0, 0, 0]);
